+++ app/api/siwe/whoami/route.ts
@@ -5,7 +5,7 @@ import { SessionData, sessionOptions } from '../session'
 
 async function handler() {  
   const session = await getIronSession<SessionData>(cookies(), sessionOptions)
-  return NextResponse.json({ address: session.siwe?.data.address })
+  return NextResponse.json({ address: session.siwe?.data.address || null })
 }
 
 export { handler as GET }
