+++ components/Signin/signals.ts
@@ -1,0 +1,10 @@
+import { signal } from 'signals-react-safe'
+
+export const siweSignal = signal({
+  signingIn: false,
+	signedIn: false
+} as {
+	signingIn: boolean
+	signedIn: boolean
+	address: string | undefined
+})
