+++ package.json
@@ -15,6 +15,7 @@
     "@yearn-finance/web-lib": "^3.0.82",
     "db-migrate": "^0.11.14",
     "db-migrate-pg": "^1.5.2",
+    "framer-motion": "^10.16.16",
     "iron-session": "^8.0.1",
     "jsonwebtoken": "^9.0.2",
     "next": "14.0.3",
@@ -22,6 +23,7 @@
     "pino-pretty": "^10.2.3",
     "react": "^18",
     "react-dom": "^18",
+    "signals-react-safe": "^1.1.2",
     "siwe": "^2.1.4",
     "viem": "^1.19.11",
     "wagmi": "1.4.7"
